+++ src/components/PreWork.tsx
@@ -1,0 +1,18 @@
+import React from "react";
+import { ContainerItem } from "./common/ContainerItem";
+import { constURL } from "../utils/ConstURL";
+import * as S from "../App.styles";
+
+export const PreWork = () => {
+  return (
+    <ContainerItem
+      title={"1. 사전 작업"}
+      content={
+        <p>
+          <S.A href={constURL.timsUrl}>TALK 로그인</S.A> (10분 전 권장) 및 해당
+          페이지 팝업 허용
+        </p>
+      }
+    />
+  );
+};
