+++ src/declare/type.d.ts
@@ -1,0 +1,3 @@
+export interface ReservationListProps {
+  reservationList: string[];
+}
